Added app name to the store
