+++ src/components/lib/VTextField/VTextField.vue
@@ -36,6 +36,10 @@ const props = defineProps({
     type: [String, Number],
     default: '',
   },
+  value: {
+    type: [String, Number],
+    default: '',
+  },
   type: {
     type: String,
     default: 'text',
@@ -74,7 +78,7 @@ const props = defineProps({
   },
 })
 
-const emit = defineEmits(['update:modelValue', 'icon-clicked'])
+const emit = defineEmits(['update:modelValue', 'icon-clicked', 'change'])
 const attrs = useAttrs()
 let componentId = props.id?.trim()
 
@@ -84,6 +88,7 @@ let inputType = computed(() => {
 
 function onInput(ev) {
   emit('update:modelValue', ev.target?.value)
+  emit('change', ev.target?.value)
 }
 
 function onIconClicked(ev) {
@@ -102,7 +107,7 @@ function onIconClicked(ev) {
       <input
         :id="componentId"
         :type="inputType"
-        :value="modelValue"
+        :value="modelValue || value"
         v-bind="attrs"
         :placeholder="placeholder"
         :class="{ strong }"
